--- crates/api/README.md
@@ -1,13 +1,0 @@
-# API
-
-## Overview
-This crate provides services for all 3 mev-boost APIs.
-
-## Features
-- **Builder API**: https://flashbots.github.io/relay-specs/#/Builder
-- **Proposer API**: https://ethereum.github.io/builder-specs/#/Builder
-- **Data API**: https://flashbots.github.io/relay-specs/#/Data
-
-## License
-
-MIT OR Apache-2.0
