+++ crates/api/src/builder/mod.rs
@@ -1,8 +1,16 @@
+#![allow(clippy::type_complexity)]
+
 pub mod api;
 pub mod error;
 pub mod simulator;
 pub mod types;
 pub mod v3;
+
+mod gossip;
+mod submit_block;
+mod submit_block_v2;
+mod submit_header;
+mod top_bid;
 
 pub use simulator::*;
 pub use types::*;
